added ability to create an orderItem, next step is creating a new order

+++ bangazoncli/Products/GetProductData.cs
@@ -27,7 +27,7 @@ namespace bangazoncli.Products
                     var product = new Product
                     {
                         Name = reader["Name"].ToString(),
-                        Price = int.Parse(reader["price"].ToString()),
+                        Price = decimal.Parse(reader["price"].ToString()),
                         ProductID = int.Parse(reader["ProductID"].ToString())
                     };
